Feature/visualize vertices

+++ mav_trajectory_generation/src/vertex.cpp
@@ -188,7 +188,7 @@ bool Vertex::getSubdimension(const std::vector<size_t>& subdimensions,
   *subvertex = Vertex(subdimensions.size());
 
   // Check if all subdimensions exist.
-  for (int subdimension : subdimensions)
+  for (size_t subdimension : subdimensions)
     if (subdimension >= D_) return false;
 
   // Copy constraints up to maximum derivative order.
@@ -263,9 +263,9 @@ std::vector<double> estimateSegmentTimesNfabian(const Vertex::Vector& vertices,
     vertices[i].getConstraint(derivative_order::POSITION, &start);
     vertices[i + 1].getConstraint(derivative_order::POSITION, &end);
     double distance = (end - start).norm();
-    double t = distance / v_max * 2 * (1.0 +
-                                       magic_fabian_constant * v_max / a_max *
-                                           exp(-distance / v_max * 2));
+    double t = distance / v_max * 2 *
+               (1.0 + magic_fabian_constant * v_max / a_max *
+                          exp(-distance / v_max * 2));
     segment_times.push_back(t);
   }
   return segment_times;
